Move tctl acl command into its own directory
