+++ tool/tctl/common/accesslist/command.go
@@ -16,7 +16,7 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
 
-package common
+package accesslist
 
 import (
 	"context"
@@ -48,8 +48,8 @@ import (
 	tctlcfg "github.com/gravitational/teleport/tool/tctl/common/config"
 )
 
-// ACLCommand implements the `tctl acl` family of commands.
-type ACLCommand struct {
+// Command implements the `tctl acl` family of commands.
+type Command struct {
 	format string
 
 	ls            *kingpin.CmdClause
@@ -86,8 +86,8 @@ const (
 	memberKindList = "list"
 )
 
-// Initialize allows ACLCommand to plug itself into the CLI parser
-func (c *ACLCommand) Initialize(app *kingpin.Application, _ *tctlcfg.GlobalCLIFlags, _ *servicecfg.Config) {
+// Initialize allows Command to plug itself into the CLI parser
+func (c *Command) Initialize(app *kingpin.Application, _ *tctlcfg.GlobalCLIFlags, _ *servicecfg.Config) {
 	acl := app.Command("acl", "Manage Access Lists.").Alias("access-lists")
 
 	c.ls = acl.Command("ls", "List cluster Access Lists.")
@@ -132,7 +132,7 @@ func (c *ACLCommand) Initialize(app *kingpin.Application, _ *tctlcfg.GlobalCLIFl
 }
 
 // TryRun takes the CLI command as an argument (like "acl ls") and executes it.
-func (c *ACLCommand) TryRun(ctx context.Context, cmd string, clientFunc commonclient.InitFunc) (match bool, err error) {
+func (c *Command) TryRun(ctx context.Context, cmd string, clientFunc commonclient.InitFunc) (match bool, err error) {
 	var commandFunc func(ctx context.Context, client *authclient.Client) error
 	switch cmd {
 	case c.ls.FullCommand():
@@ -163,7 +163,7 @@ func (c *ACLCommand) TryRun(ctx context.Context, cmd string, clientFunc commoncl
 }
 
 // List will list access lists visible to the user.
-func (c *ACLCommand) List(ctx context.Context, client *authclient.Client) error {
+func (c *Command) List(ctx context.Context, client *authclient.Client) error {
 	var accessLists []*accesslist.AccessList
 	var err error
 
@@ -214,12 +214,12 @@ func parseACLQualifiedName(name string) (scopes.QualifiedName, error) {
 	return scopes.QualifiedName{Name: name}, nil
 }
 
-func (c *ACLCommand) accessListQualifiedName() (scopes.QualifiedName, error) {
+func (c *Command) accessListQualifiedName() (scopes.QualifiedName, error) {
 	sqn, err := parseACLQualifiedName(c.accessListName)
 	return sqn, trace.Wrap(err, "validating access list name as a scope-qualified name")
 }
 
-func (c *ACLCommand) memberQualifiedName() (scopes.QualifiedName, error) {
+func (c *Command) memberQualifiedName() (scopes.QualifiedName, error) {
 	sqn, err := parseACLQualifiedName(c.memberName)
 	return sqn, trace.Wrap(err, "validating member name as a scope-qualified name")
 }
@@ -237,7 +237,7 @@ func splitACLQualifiedNames(names string) ([]scopes.QualifiedName, error) {
 }
 
 // Get will display information about an access list visible to the user.
-func (c *ACLCommand) Get(ctx context.Context, client *authclient.Client) error {
+func (c *Command) Get(ctx context.Context, client *authclient.Client) error {
 	aclName, err := c.accessListQualifiedName()
 	if err != nil {
 		return trace.Wrap(err)
@@ -254,7 +254,7 @@ func (c *ACLCommand) Get(ctx context.Context, client *authclient.Client) error {
 }
 
 // UsersAdd will add a user to an access list.
-func (c *ACLCommand) UsersAdd(ctx context.Context, client *authclient.Client) error {
+func (c *Command) UsersAdd(ctx context.Context, client *authclient.Client) error {
 	var expires time.Time
 	if c.expires != "" {
 		var err error
@@ -314,7 +314,7 @@ func (c *ACLCommand) UsersAdd(ctx context.Context, client *authclient.Client) er
 }
 
 // UsersRemove will remove a user to an access list.
-func (c *ACLCommand) UsersRemove(ctx context.Context, client *authclient.Client) error {
+func (c *Command) UsersRemove(ctx context.Context, client *authclient.Client) error {
 	aclName, err := c.accessListQualifiedName()
 	if err != nil {
 		return trace.Wrap(err)
@@ -339,7 +339,7 @@ func (c *ACLCommand) UsersRemove(ctx context.Context, client *authclient.Client)
 }
 
 // UsersList will list the users in an access list.
-func (c *ACLCommand) UsersList(ctx context.Context, client *authclient.Client) error {
+func (c *Command) UsersList(ctx context.Context, client *authclient.Client) error {
 	aclName, err := c.accessListQualifiedName()
 	if err != nil {
 		return trace.Wrap(err)
@@ -392,7 +392,7 @@ func (c *ACLCommand) UsersList(ctx context.Context, client *authclient.Client) e
 	}
 }
 
-func (c *ACLCommand) ReviewsCreate(ctx context.Context, client *authclient.Client) error {
+func (c *Command) ReviewsCreate(ctx context.Context, client *authclient.Client) error {
 	review, err := c.makeReview()
 	if err != nil {
 		return trace.Wrap(err)
@@ -408,7 +408,7 @@ func (c *ACLCommand) ReviewsCreate(ctx context.Context, client *authclient.Clien
 	return nil
 }
 
-func (c *ACLCommand) makeReview() (*accesslist.Review, error) {
+func (c *Command) makeReview() (*accesslist.Review, error) {
 	aclName, err := c.accessListQualifiedName()
 	if err != nil {
 		return nil, trace.Wrap(err)
@@ -443,7 +443,7 @@ func (c *ACLCommand) makeReview() (*accesslist.Review, error) {
 		aclName.Scope)
 }
 
-func (c *ACLCommand) ReviewsList(ctx context.Context, client *authclient.Client) error {
+func (c *Command) ReviewsList(ctx context.Context, client *authclient.Client) error {
 	aclName, err := c.accessListQualifiedName()
 	if err != nil {
 		return trace.Wrap(err)
@@ -466,7 +466,7 @@ func (c *ACLCommand) ReviewsList(ctx context.Context, client *authclient.Client)
 	return trace.Wrap(c.displayAccessListReviews(reviews))
 }
 
-func (c *ACLCommand) displayAccessListReviews(reviews []*accesslist.Review) error {
+func (c *Command) displayAccessListReviews(reviews []*accesslist.Review) error {
 	switch c.format {
 	case teleport.YAML:
 		return trace.Wrap(utils.WriteYAML(c.Stdout, reviews))
@@ -478,7 +478,7 @@ func (c *ACLCommand) displayAccessListReviews(reviews []*accesslist.Review) erro
 	return trace.BadParameter("invalid format %q", c.format)
 }
 
-func (c *ACLCommand) displayAccessListReviewsText(reviews []*accesslist.Review) error {
+func (c *Command) displayAccessListReviewsText(reviews []*accesslist.Review) error {
 	table := asciitable.MakeTable([]string{"ID", "Reviewer", "Review Date", "Removed Members", "Notes"})
 	for _, review := range reviews {
 		table.AddRow([]string{
@@ -493,7 +493,7 @@ func (c *ACLCommand) displayAccessListReviewsText(reviews []*accesslist.Review) 
 	return trace.Wrap(err)
 }
 
-func (c *ACLCommand) displayAccessLists(accessLists ...*accesslist.AccessList) error {
+func (c *Command) displayAccessLists(accessLists ...*accesslist.AccessList) error {
 	switch c.format {
 	case teleport.YAML:
 		return trace.Wrap(utils.WriteYAML(c.Stdout, accessLists))
@@ -507,7 +507,7 @@ func (c *ACLCommand) displayAccessLists(accessLists ...*accesslist.AccessList) e
 	return trace.BadParameter("invalid format %q", c.format)
 }
 
-func (c *ACLCommand) displayAccessListsText(accessLists ...*accesslist.AccessList) error {
+func (c *Command) displayAccessListsText(accessLists ...*accesslist.AccessList) error {
 	table := asciitable.MakeTable([]string{"ID", "Title", "Next Audit", "Granted Roles", "Granted Traits"})
 	for _, accessList := range accessLists {
 		grantedRoles := append([]string(nil), accessList.GetGrants().Roles...)
