+++ tool/tctl/common/cmds.go
@@ -19,6 +19,7 @@
 package common
 
 import (
+	"github.com/gravitational/teleport/tool/tctl/common/accesslist"
 	"github.com/gravitational/teleport/tool/tctl/common/accessmonitoring"
 	"github.com/gravitational/teleport/tool/tctl/common/decision"
 	"github.com/gravitational/teleport/tool/tctl/common/discovery"
@@ -60,7 +61,7 @@ func Commands() []CLICommand {
 		&LoadtestCommand{},
 		&DevicesCommand{},
 		&SAMLCommand{},
-		&ACLCommand{},
+		&accesslist.Command{},
 		&loginrule.Command{},
 		&IdPCommand{},
 		&accessmonitoring.Command{},
